moved coordinaterecorder within timecontroller plugin
